+++ src/main/scala/ai/starlake/config/Settings.scala
@@ -594,7 +594,7 @@ object Settings extends LazyLogging {
 
     @JsonIgnore
     lazy val dialect: JdbcDialect =
-      applyIfConnectionTypeIs(ConnectionType.JDBC, SparkUtils.dialect(jdbcUrl))
+      applyIfConnectionTypeIs(ConnectionType.JDBC, SparkUtils.dialectForUrl(jdbcUrl))
 
     def quoteIdentifier(identifier: String): String = dialect.quoteIdentifier(identifier)
 
@@ -1347,8 +1347,21 @@ object Settings extends LazyLogging {
     // Load fairscheduler.xml
     val jobConf = initSparkConfig(applicationConfSettings)
     val withSparkConfig = applicationConfSettings.copy(jobConf = jobConf)
-    val withDefaultSchdules = addDefaultSchedules(withSparkConfig)
-    withDefaultSchdules
+    val withDefaultSchedules = addDefaultSchedules(withSparkConfig)
+
+    val localCatalog = jobConf.get("spark.localCatalog", "none")
+    localCatalog match {
+      case "none"              => // do nothing
+      case "iceberg" | "delta" =>
+
+    }
+
+    val withDefaultWriteFormat =
+      withDefaultSchedules.copy(appConfig =
+        withDefaultSchedules.appConfig.copy(defaultWriteFormat = localCatalog)
+      )
+
+    withDefaultWriteFormat
   }
 
   val defaultCronPresets = Map(